Structural changes and cleanup added utils.c/.h files with general purpose functions cleaned up code and added some comments
