+++ .vscode/settings.json
@@ -18,6 +18,9 @@
         "functional": "c",
         "tuple": "c",
         "utility": "c",
-        "unistd.h": "c"
+        "unistd.h": "c",
+        "util.h": "c",
+        "stdlib.h": "c",
+        "utils.h": "c"
     }
 }
