+++ include/hash1.h
@@ -4,4 +4,3 @@
 
 
 unsigned int hash1(unsigned int , int );
-relation* partition_relation(relation ,int );
