+++ include/partition.h
@@ -1,3 +1,10 @@
+#include <stdio.h>
+#include <stdlib.h>
+#include <unistd.h>
+#include <math.h>
+
+#include "hash1.h"
+#include "utils.h"
 #include "relations.h"
 
-relation* partition_relation(relation,int);
+void partition_relation(relation,int);
